+++ alicloud/data_source_alicloud_zones.go
@@ -154,7 +154,7 @@ func dataSourceAlicloudZonesRead(d *schema.ResourceData, meta interface{}) error
 			return fmt.Errorf("[ERROR] There is no available region for KVStore")
 		}
 		for _, r := range regions.RegionIds.KVStoreRegion {
-			for _, zoneID := range strings.Split(r.ZoneIds, ",") {
+			for _, zoneID := range r.ZoneIdList.ZoneId {
 				if multi && strings.Contains(zoneID, MULTI_IZ_SYMBOL) && r.RegionId == string(meta.(*aliyunclient.AliyunClient).Region) {
 					zoneIds = append(zoneIds, zoneID)
 					continue